cambio Importe = float64

+++ servicios/novedades/novedades.go
@@ -92,7 +92,7 @@ type Distribuciones struct {
 }
 
 type RecursosNovedades struct {
-	Importe     int           `bson:"importe,truncate"`
+	Importe     float64       `bson:"importe,truncate"`
 	Comentarios string        `bson:"comentarios"`
 	Recurso     string        `bson:"recurso"`
 	Periodo     string        `bson:"periodo"`
